+++ ChangeLog.md
@@ -7,6 +7,8 @@ This project's release branch is `master`. This log is written from the perspect
 * Fix crashes of Android apps on 32-bit ARM devices.
 * Provide a way to indicate acceptance of the Android SDK license by passing `config.android_sdk.accept_license = true;` in the arguments to the import of `.obelisk/impl` in the project's `default.nix`.
 * Add `COMPLETE` pragma to `(:/)`. Using this pattern synonym should no longer generate spurious warnings about non-exhaustive pattern matching.
+* Make asset path hashing strict (see `Obelisk.Asset.Gather`)
+* Add the `ob shell` command to enter a nix shell for an obelisk project
 
 ## v0.1.0.0 - 2019-03-29